chore(ui-kit): rename files for clarity; formatted files
